requirements.txt
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
mlflow>=2.16.0
pyyaml>=6.0.0